pyproject.toml
[project]
name = "light-calc-page"
version = "0.1.0"
description = ""
authors = [
    {name = "Manabu-Aihara",email = "manabu@yoboiryo.co.jp"}
]
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "cryptography (>=44.0.1,<45.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
